refactor(tests): reorganize assets into centralized structure
- Move integration_test_data/* to tests/assets/scenarios/license_scenarios/
- Move scattered assets to tests/assets/samples/pickles/
- Rename evil_pickle.py to generate_evil_pickle.py in tests/assets/generators/
- Update all test file imports and path references

All assets properly moved as renames (not deletions) and all tests passing.
