+++ tests/assets/generators/generate_evil_pickle.py
