+++ tests/assets/samples/pickles/dill_func.pkl
