+++ tests/assets/samples/pickles/evil.pickle
