+++ tests/assets/scenarios/license_scenarios/agpl_component/agpl_model.pkl
