+++ tests/assets/scenarios/license_scenarios/agpl_component/neural_network.py
