+++ tests/assets/scenarios/license_scenarios/mit_model/config.json
