+++ tests/assets/scenarios/license_scenarios/mit_model/model_weights.pkl
