+++ tests/assets/scenarios/license_scenarios/mixed_licenses/LICENSE
