+++ tests/assets/scenarios/license_scenarios/mixed_licenses/apache_component.py
