+++ tests/assets/scenarios/license_scenarios/mixed_licenses/dataset_cc_nc.json
