+++ tests/assets/scenarios/license_scenarios/mixed_licenses/gpl_utility.py
