+++ tests/assets/scenarios/license_scenarios/unlicensed_dataset/embeddings.npy
